pyproject.toml
[project]
name = "tour-planner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai[tools]>=0.121.0",
    "fastapi>=0.115.12",
    "firecrawl-py>=2.7.0",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-core>=0.3.61",
    "langchain-groq>=0.3.2",
    "mcp-use>=1.3.0",
    "mcp[cli]>=1.9.3",
    "outlines>=0.2.3",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "streamlit>=1.45.1",
    "tools>=1.0.2",
    "unstructured>=0.17.2",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
